Added BMI160 options and schematics
I'm not entirely sure if it's okay to keep adding additional options the way I did in schematics.js (Or if it works the way I have it set). If a few more schematics get added, I think it would probably make more sense to have a switch statement instead.