Changed 4+1 battery link to 5 battery link.
Old link (https://www.aliexpress.com/item/33021202630.html) was broken and did not lead anywhere, and I just so happened to find a 5 battery set of exactly the same type of batteries (EHAO 804040) as the set of 10 for probably the same price? I am unsure whether to include delivery price or current discount, but basic price is $15 currently.