MPU+QMC Details, BMI Fix, and Defines.h Generator Update
Added MPU+QMC5883L tracker schematic, as well as noted instructions for flashing firmware.

The BMI160 schematic previously had an extraneous detail, which was corrected.

The defines.h generator has been updated to allow selecting the BMI160 and ICM20948 as options. As well, the note that the MPU9250 is experimental has been removed and replaced with a noticed about requiring a stable magnetic environment. Likewise, I attempted to a similar warning for the ICM20948 which notes that to run it in 9DoF mode requires changing an additional setting in debug.h, with another warning about requiring a stable magnetic environment.