Minor Update for Calibration Information
Despite information on calibration being included in the Docs, some people are still unsure of where to find this information or are unaware that their IMU needs to be calibrated at all! As such, an additional note on calibration is included on the IMU Calibration page with reference to where calibration instructions can be found.