Component Breakdown Rewrite
Component breakdown has been significantly rewritten. Due to the significant changes to the price calculator it no longer made sense to advertise the same links that can be seen in the Quick Link area when options can now substantially vary. Similarly, the component breakdown has shifted away from explaining the individual parts chosen, i.e. "such and such AliExpress link is this or that", and instead gives a more comprehensive view on what the components are for, how they're used, and any additional information that might be useful to people looking to DIY.