Updated Survey (Cleared Previous Results)
Given the number of old results potentially tainting the results, it is probably for the best that the survey be periodically be reset.

The results of the survey can still be accessed at the old link. If anyone wants the old results shared with them, message me on Discord (NWB#5135)