chore: update ShadowPC Documentation
Summary:
The ShadowPC client recently implemented filtering to what types of USB devices are allowed to be forwarded with the official client and now no longer allows forwarding of Network Cards. This updates the documentation to more closely reflect the actual usage the original user used when they used the ShadowPC with their DIY Slimes. It now uses the Virtual Here exclusively in the documentation since based on other Discord user troubleshooting that method remains operational like it was when originally using this approach to enjoy SlimeVR on a ShadowPC.